implement saving throws per class level
