+++ osrlib/osrlib/__init__.py
@@ -47,6 +47,9 @@ from .dungeon import (
     LocationNotFoundError,
     Direction,
 )
+from .enums import (
+    CharacterClassType,
+)
 from .game_manager import (
     GameManager,
     StorageType,
@@ -89,4 +92,5 @@ from .party import (
 )
 from .player_character import PlayerCharacter, Alignment
 from .quest import Quest
+from .saving_throws import saving_throws
 from .utils import format_modifiers
