+++ osrlib/osrlib/enums.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+class CharacterClassType(Enum):
+    """Enum representing the types of character classes."""
+
+    CLERIC = "Cleric"
+    DWARF = "Dwarf"
+    ELF = "Elf"
+    FIGHTER = "Fighter"
+    HALFLING = "Halfling"
+    MAGIC_USER = "Magic User"
+    THIEF = "Thief"
+    COMMONER = "Commoner"
